requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
PyMuPDF
tqdm
pandas
numpy
jsonlines